+++ demo/src/main/java/com/inventario/demo/entities/tenant/mapper/TenantMapper.java
@@ -1,0 +1,26 @@
+package com.inventario.demo.entities.tenant.mapper;
+
+import com.inventario.demo.entities.tenant.model.TenantModel;
+import com.inventario.demo.entities.tenant.dtoRequest.TenantRequestDto;
+import com.inventario.demo.entities.tenant.dtoResponse.TenantResponseDto;
+import org.modelmapper.ModelMapper;
+
+public class TenantMapper {
+    private final ModelMapper modelMapper;
+
+    public TenantMapper(ModelMapper modelMapper) {
+        this.modelMapper = modelMapper;
+    }
+
+    public TenantResponseDto toDto(TenantModel tenant) {
+        return modelMapper.map(tenant, TenantResponseDto.class);
+    }
+
+    public TenantModel toEntity(TenantRequestDto tenantRequestDto) {
+        return modelMapper.map(tenantRequestDto, TenantModel.class);
+    }
+
+    public void updateEntityFromDto(TenantRequestDto dto, TenantModel entity) {
+        modelMapper.map(dto, entity);
+    }
+}
