+++ demo/src/main/java/com/inventario/demo/entities/user/controller/UserController.java
@@ -3,9 +3,8 @@ package com.inventario.demo.entities.user.controller;
 import com.inventario.demo.entities.user.dtoRequest.UserRequestDto;
 import com.inventario.demo.entities.user.dtoResponse.UserPageResponse;
 import com.inventario.demo.entities.user.dtoResponse.UserResponseDto;
-import com.inventario.demo.entities.user.model.UserModel;
 import com.inventario.demo.entities.user.repository.UserRepository;
-import com.inventario.demo.entities.user.service.UserService;
+import com.inventario.demo.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
