--- demo/src/main/java/com/inventario/demo/entities/user/model/UserModel.java~
@@ -1,62 +1,0 @@
-package com.inventario.demo.user.model;
-
-import com.inventario.demo.tenant.model.TenantModel;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDate;
-import java.util.HashSet;
-import java.util.Set;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Table(name = "users")
-public class UserModel {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    private String name;
-
-    @Column(name = "last_name")
-    private String lastName;
-
-    @Column(name = "phone_number")
-    private Long phoneNumber;
-
-    private String country;
-
-    @Column(name = "birth_date")
-    private LocalDate birthDate;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "tenant_id", nullable = false)
-    private TenantModel tenant;
-
-    @CreationTimestamp
-    @Column(name = "register_date", updatable = false)
-    private LocalDate registerDate;
-
-    @UpdateTimestamp
-    @Column(name = "last_login")
-    private LocalDate lastLogin;
-
-    @ManyToMany(fetch = FetchType.LAZY)
-    @JoinTable(name = "users_roles",
-            joinColumns = @JoinColumn(name = "user_id"),
-            inverseJoinColumns = @JoinColumn(name = "role_id"))
-    private Set<RoleModel> roles = new HashSet<>();
-
-}
