--- demo/src/main/java/com/inventario/demo/tenant/model/TenantModel.java
@@ -1,42 +1,0 @@
-package com.inventario.demo.tenant.model;
-
-import com.vladmihalcea.hibernate.type.json.JsonBinaryType;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.Type;
-import org.hibernate.annotations.UpdateTimestamp;
-
-
-import java.time.LocalDate;
-import java.util.HashMap;
-import java.util.Map;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Table(name = "tenants")
-public class TenantModel {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Type(value = JsonBinaryType.class)
-    @Column(columnDefinition = "jsonb")
-    private Map<String, Object> configuration = new HashMap<>();
-
-    @CreationTimestamp
-    @Column(name = "created_at", nullable = false, updatable = false)
-    private LocalDate createdAt;
-
-    @UpdateTimestamp
-    @Column(name = "updated_at", nullable = false)
-    private LocalDate updatedAt;
-
-}
