--- demo/src/main/java/com/inventario/demo/tenant/repository/TenantRepository.java
@@ -1,8 +1,0 @@
-package com.inventario.demo.tenant.repository;
-
-import com.inventario.demo.tenant.model.TenantModel;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TenantRepository extends JpaRepository<TenantModel, Long> {
-
-}
