--- demo/src/main/java/com/inventario/demo/transaction/respository/TransactionRepository.java
@@ -1,7 +1,0 @@
-package com.inventario.demo.transaction.respository;
-
-import com.inventario.demo.transaction.model.TransactionModel;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TransactionRepository extends JpaRepository<TransactionModel, Long> {
-}
